Extremely rudimentary functionality working
A player is added via mapEntranceNormal. Nothing happens outside of
graphics drawing, but that 's more than enough to warrant a commit.

Palette generation added to PixelRender and given a UI in ImageReadr.

More documentation fixups.

More genericism. A lot of these are being pushed to FullScreenMario -
when the repositories are more stable, GameStartr should be given its
own repo and git-linked to these two (as should its own modules).