Refactored MenuGraphr::addMenuWord for new string[]s
This was required for first-class support of string[]-stored names. The
code is a lot cleaner now.

Instead of directly allowing the various types of dialogs in
addMenuWord, they're instead only allowed in addMenuDialog, and then
parsed into the standard format.

The standard format is to have an array of lines, where a line is a
mixed array of words or command objects; words are each string[]s.