Fixed cycling going through walls
Flying through the air at hundreds of miles an hour,  with an altitude
of 37,005 feet, it is finally fixed.

Borders needed to be accounted for more accurately. Lower speeds were
fine because the player would still be found as within the region after
overlapping, but higher speeds let the player just blow past the guards.

Fixes #178