Added AreaGates to help spawn across areas (#350)
* Added an AreaGate detector

It changes the current area/map/location, but doesn't mess with
offsetes.

* Fixed AreaSpawner direction-based size glitches

* Made ItemsHoldr source-of-truth for map/area/loc

The previous system had it set in saveGame, which meant whatever was
changed in AreaGate activations would be overriden. Now,
`FSP.maps.setLocation` stores it in ItemsHoldr (excluding the Blank
map).

* Added logic to move MapScreenr on area change

The logic is slightly flawed at the moment. Will file a separate issue
to fix it.