Complete battles refactor (#403)
* Started bringing in refactored BattleMovr

Wiped out most code that deals with battles, either by real deletion or
commenting out. Clean slate!

* Started clearing battle transitions, animations

* Partially fleshed out player battle selector

* Fixed up battle options to be mostly optional

* Started on battle visuals

* Added visual setup for Things

* Reworked battles through the player selector

Got rid of the dependency on ScenePlayr (filed #396), and moved to using
animation components.

* Added player Pokemon entrance animations

Reworked battle animations to be more componentized

* Started on the player's health display

It's a bit offset, but it's a start.

* Fixed the initial opponent battle health display

* Fleshed out pokeball displays pre-health

* Fixed up the stats menu

* Fixed sizes for secondary stats menu

* Added ActionsOrderer for battles

Still needs to have the opponent's actions added.

* Fleshed out the OpponentSelector

It (mostly correctly) takes into accord the standard RBY equations.

A couple of major todos:
* Items, in general
* Effects for moves other than Growl will need to be filled out.
* Lorelei!

* Moved EffectTarget, IEffectBase to constants

* Got repeatable action selection to work

Constants for battles are now split into Modifications and Texts.

* Brought in BattleMovr's ITeamAndAction

* Added player fleeing

It doesn't actually work beyond calling BattleMovr's stopBattle, mind
you. The onComplete callback for that still needs to be implemented.

* Split ending animations into Ending class

* Added Pokemon switching

* Switched battle texts to be generators

* Genericized switching animations

Added a new "shared" directory with settings that the teams' animators
can implement.

* Moved all sub-animations to shared

Also added the initial Damage, Missed, Statistics, Statuses, and
Switching skeletons.

* Fixed Effects to actually call onComplete

Also added Scratch, Tackle effects

* Added health bar changing for damage

* Added fainting for switching

* Commented GrowlMove.ts

* Added most of the previously implemented moves

Skipping a couple because their sprites are missing.

* Fixed missing type in Battles component

* Fixed casing for Battles texts import

* Removed [i: number] from raw maps

* Re-added number keys for raw maps

`src:tsc` fails for this on Travis but not locally...