Re-added the Oak lab rival battle (#419)
* Fixed battles' keptThings to be behind backgrounds

The white background text Thing was behing placed first *after* they
were already moved to first, resulting in keptThings being on top of it
during battle. Super awkward.

* Fixed CHARMANDERFront sprite crashes

Another case of the sprite being too big.

* Re-added OakIntroRivalBattleCutscene's battle

Some texts are wrong, but otherwise it starts

* Fixed bug with attackingTeam move choices

It was passing the move object instead of the move name. Result of
https://github.com/FullScreenShenanigans/NumberMakr/issues/6.

Also fixed OakIntroRivalBattleCutscene to define the leader.

* Fixed up battle onComplete piping

* Fixed routine names for OakIntroRivalLeavesCutscene

They all had a "Leaves" prefix, which shouldn't have been there.

* Re-added in-battle ending dialogs

* Fixed casing for battle animations

* Moved afterBattle to end of IBattleTextGenerators

Related to one of the open TSLint bugs with `member-ordering`.