Switched to userwrappr@0.7.2, gamestartr@0.7.3 (#532)
Constructor takes in size settings instead of waiting to reset. This way, the game is always guaranteed to have modules & components - which is less confusing for the type system. UserWrappr logic needs mod names before instantiating the game, so the names are now static members of the `ModComponent` classes accessed with a getter. An `InterfaceFactory` instantiates the UserWrappr logic.