Added tests for PokeCenter healing
Moved the internals of PokeCenterCutscene off of the old ScenePlayr routine structure. They're just regular callbacks now.
The tests themselves are a bit unweildy: they do fancy math to try to anticipate how long things take in the cutscene. I can't think of a better way for now.

As a workaround to make the tests not take 500ms, set `framerateSkip` in the stubbed games to `9000001`.  Stops the game from redrawing on subsequent ticks.